+++ internal/consts.go
@@ -1,0 +1,8 @@
+package internal
+
+import (
+	"errors"
+)
+
+var ErrFlagNotFound = errors.New("flag not found")
+var ErrFlagNotType = errors.New("flag is not of the expected type")